refactor: 플레이리스트 이미지 저장 메서드 뷰모델로 이동 (#237)
- 알럿 처리 메서드로 정리
